test(cli): добавлены недостающие тесты для create-federation-plugin
